api/requirements.txt
Flask==2.2.3
Werkzeug==2.2.3
requests==2.32.3
google-cloud-firestore==2.7.0
gunicorn
python-dotenv
jinja2
scikit-learn
google-cloud-storage>=2.0.0


